moved draftjs to android module
