+++ settings.gradle
@@ -31,3 +31,4 @@ include ':feature-search' // Feature module for Search
 include ':feature-bookmark' // Feature module for Bookmark
 include ':capabilities' // Capabilities module
 include ':core' // Core module
+include ':draftjs-compose' // Draftjs compose module
